+++ src/components/input/Input.vue
@@ -10,6 +10,7 @@
 			v-bind="$attrs"
 			:disabled="disabled"
 			:type="type"
+			:value="currentValue"
 			@input="onInput"
 			@change="onChange"
 			@focus="onFocus"
@@ -37,6 +38,27 @@
 				</svg>
 			</span>
 		</span>
+		<span
+			v-if="(clearable && currentValue && !disabled) || indicator"
+			ref="indicator"
+			class="wvui-input__indicator"
+			@click="onClear"
+		>
+			<span class="wvui-icon">
+				<svg
+					xmlns="http://www.w3.org/2000/svg"
+					width="12"
+					height="12"
+					viewBox="0 0 20 20"
+					aria-hidden="true"
+					role="presentation"
+				>
+					<g fill="#72777d">
+						<path :d="clearIcon" />
+					</g>
+				</svg>
+			</span>
+		</span>
 	</div>
 </template>
 
@@ -47,11 +69,21 @@ import Vue, { PropType } from 'vue';
 // Hardcoded svg for search icon, will be removed after icon <wvui-icon/> is ready
 // eslint-disable-next-line max-len
 const mwIconSearch = 'M7.5 13c3.04 0 5.5-2.46 5.5-5.5S10.54 2 7.5 2 2 4.46 2 7.5 4.46 13 7.5 13zm4.55.46A7.432 7.432 0 0 1 7.5 15C3.36 15 0 11.64 0 7.5S3.36 0 7.5 0C11.64 0 15 3.36 15 7.5c0 1.71-.57 3.29-1.54 4.55l6.49 6.49-1.41 1.41-6.49-6.49z';
+// Hardcoded svg for clear action, will be removed after icon <wvui-icon/> is ready
+// eslint-disable-next-line max-len
+const mwIconClose = 'M4.34 2.93l12.73 12.73-1.41 1.41L2.93 4.35z M17.07 4.34L4.34 17.07l-1.41-1.41L15.66 2.93z';
+// Hardcoded svg for indicator , will be removed after icon <wvui-icon/> is ready
+// eslint-disable-next-line max-len
+export const mwIconInfo = 'M9.5 16A6.61 6.61 0 0 1 3 9.5 6.61 6.61 0 0 1 9.5 3 6.61 6.61 0 0 1 16 9.5 6.63 6.63 0 0 1 9.5 16zm0-14A7.5 7.5 0 1 0 17 9.5 7.5 7.5 0 0 0 9.5 2zm.5 6v4.08h1V13H8.07v-.92H9V9H8V8zM9 6h1v1H9z';
 
 export default Vue.extend( {
 	name: 'WvuiInput',
 	inheritAttrs: false,
 	props: {
+		value: {
+			type: [ String, Number ],
+			default: ''
+		},
 		type: {
 			type: String as PropType<InputType>,
 			default: InputType.Text,
@@ -68,25 +100,40 @@ export default Vue.extend( {
 		indicator: {
 			type: String,
 			default: null
+		},
+		// clearable properyy will override indicator property
+		clearable: {
+			type: Boolean,
+			default: false
 		}
 	},
 	data() {
 		return {
+			currentValue: this.value,
 			// temporary hardcoded icons
-			searchIcon: mwIconSearch
+			searchIcon: mwIconSearch,
+			clearIcon: this.clearable ? mwIconClose : mwIconInfo
 		};
 
 	},
 	computed: {
 		rootClasses(): Record<string, boolean> {
 			return {
-				'wvui-input--icon': !!this.icon
+				'wvui-input--has-icon': !!this.icon,
+				'wvui-input--clearable': this.clearable
 			};
+		},
+		iconName(): string {
+			return this.clearable ? 'clear' : this.icon;
 		}
 	},
 	methods: {
 		onInput( event: InputEvent ): void {
-			this.$emit( 'input', event );
+			const target = event.target as HTMLInputElement;
+			const { value } = target;
+
+			this.setCurrentValue( value );
+			this.$emit( 'input', value );
 		},
 		onChange( event: Event ): void {
 			this.$emit( 'change', event );
@@ -96,6 +143,17 @@ export default Vue.extend( {
 		},
 		onBlur( event: FocusEvent ): void {
 			this.$emit( 'blur', event );
+		},
+		onClear(): void {
+			if ( !this.clearable ) {
+				return;
+			}
+
+			this.$emit( 'input', '' );
+			this.setCurrentValue( '' );
+		},
+		setCurrentValue( value: string | number ): void {
+			this.currentValue = value;
 		}
 	}
 } );
@@ -109,13 +167,29 @@ export default Vue.extend( {
 	vertical-align: middle;
 	box-sizing: border-box;
 
-	&__icon {
+	&__icon,
+	&__indicator {
 		position: absolute;
 		top: 50%;
 		transform: translateY( -50% );
 		line-height: 1;
 		padding-left: @padding-horizontal-input-text;
 		pointer-events: none;
+	}
+
+	&__icon {
+		padding-left: @padding-horizontal-input-text;
+	}
+
+	&__indicator {
+		right: 0;
+		padding-right: @padding-horizontal-input-text;
+	}
+
+	&--clearable {
+		.wvui-input__indicator {
+			cursor: pointer;
+		}
 	}
 
 	&__input {
@@ -178,7 +252,7 @@ export default Vue.extend( {
 		}
 	}
 
-	&--icon {
+	&--has-icon {
 		.wvui-input__input {
 			padding-left: @padding-horizontal-input-text * 2 + @size-icon;
 		}